Simplify adding common files to projects
