+++ src/System/Net/Benchmarks/Apps/Tls/Quic/Client/QuicClient.csproj
@@ -7,33 +7,18 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <ServerGarbageCollection>true</ServerGarbageCollection>
     <EnablePreviewFeatures>true</EnablePreviewFeatures>
-    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
     <NoWarn>$(NoWarn);CA1416</NoWarn> <!-- "This call site is reachable on all platforms. It is only supported on: 'linux', 'macOS/OSX', 'windows'." -->
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="Program.cs" />
-    <Compile Include="QuicClientConnection.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkApp.cs"
-             Link="Common/BenchmarkApp.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkClient.cs"
-             Link="Common/BenchmarkClient.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkOptionsBinder.cs"
-             Link="Common/BenchmarkOptionsBinder.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)LogHelper.cs"
-             Link="Common/LogHelper.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/ApplicationProtocolConstants.cs"
-             Link="Common/Tls/ApplicationProtocolConstants.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkClient.cs"
-             Link="Common/Tls/TlsBenchmarkClient.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkClientOptions.cs"
-             Link="Common/Tls/TlsBenchmarkClientOptions.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkClientOptionsBinder.cs"
-             Link="Common/Tls/TlsBenchmarkClientOptionsBinder.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkOptions.cs"
-             Link="Common/Tls/TlsBenchmarkOptions.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkOptionsHelper.cs"
-             Link="Common/Tls/TlsBenchmarkOptionsHelper.cs" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Client/*.cs"
+             LinkBase="Common" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Shared/*.cs"
+             LinkBase="Common" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/Client/*.cs"
+             LinkBase="Common/Tls" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/Shared/*.cs"
+             LinkBase="Common/Tls" />
   </ItemGroup>
 
   <ItemGroup>
