+++ src/System/Net/Benchmarks/Apps/Tls/Quic/Server/QuicServer.csproj
@@ -7,34 +7,18 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <ServerGarbageCollection>true</ServerGarbageCollection>
     <EnablePreviewFeatures>true</EnablePreviewFeatures>
-    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
     <NoWarn>$(NoWarn);CA1416</NoWarn> <!-- "This call site is reachable on all platforms. It is only supported on: 'linux', 'macOS/OSX', 'windows'." -->
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="Program.cs" />
-    <Compile Include="QuicServerConnection.cs" />
-    <Compile Include="QuicServerListener.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkApp.cs"
-             Link="Common/BenchmarkApp.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkOptionsBinder.cs"
-             Link="Common/BenchmarkOptionsBinder.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkServer.cs"
-             Link="Common/BenchmarkServer.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)LogHelper.cs"
-             Link="Common/LogHelper.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/ApplicationProtocolConstants.cs"
-             Link="Common/Tls/ApplicationProtocolConstants.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkServer.cs"
-             Link="Common/Tls/TlsBenchmarkServer.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkServerOptions.cs"
-             Link="Common/Tls/TlsBenchmarkServerOptions.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkServerOptionsBinder.cs"
-             Link="Common/Tls/TlsBenchmarkServerOptionsBinder.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkOptions.cs"
-             Link="Common/Tls/TlsBenchmarkOptions.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkOptionsHelper.cs"
-             Link="Common/Tls/TlsBenchmarkOptionsHelper.cs" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Server/*.cs"
+             LinkBase="Common" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Shared/*.cs"
+             LinkBase="Common" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/Server/*.cs"
+             LinkBase="Common/Tls" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/Shared/*.cs"
+             LinkBase="Common/Tls" />
   </ItemGroup>
 
   <ItemGroup>
