+++ src/System/Net/Benchmarks/Apps/Tls/SslStream/Client/SslStreamClient.csproj
@@ -6,36 +6,18 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <ServerGarbageCollection>true</ServerGarbageCollection>
-    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="Program.cs" />
-    <Compile Include="SslStreamClientConnection.cs" />
-    <Compile Include="SslStreamClientOptions.cs" />
-    <Compile Include="SslStreamClientOptionsBinder.cs" />
-    <Compile Include="../Shared/SslStreamOptionsHelper.cs"
-             Link="SslStream/Shared/SslStreamOptionsHelper.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkApp.cs"
-             Link="Common/BenchmarkApp.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkClient.cs"
-             Link="Common/BenchmarkClient.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)BenchmarkOptionsBinder.cs"
-             Link="Common/BenchmarkOptionsBinder.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)LogHelper.cs"
-             Link="Common/LogHelper.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/ApplicationProtocolConstants.cs"
-             Link="Common/Tls/ApplicationProtocolConstants.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkClient.cs"
-             Link="Common/Tls/TlsBenchmarkClient.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkClientOptions.cs"
-             Link="Common/Tls/TlsBenchmarkClientOptions.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkClientOptionsBinder.cs"
-             Link="Common/Tls/TlsBenchmarkClientOptionsBinder.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkOptions.cs"
-             Link="Common/Tls/TlsBenchmarkOptions.cs" />
-    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/TlsBenchmarkOptionsHelper.cs"
-             Link="Common/Tls/TlsBenchmarkOptionsHelper.cs" />
+    <Compile Include="../Shared/*.cs" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Client/*.cs"
+             LinkBase="Common" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Shared/*.cs"
+             LinkBase="Common" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/Client/*.cs"
+             LinkBase="Common/Tls" />
+    <Compile Include="$(SystemNetBenchmarksCommonPath)Tls/Shared/*.cs"
+             LinkBase="Common/Tls" />
   </ItemGroup>
 
   <ItemGroup>
