+++ src/System/Net/Benchmarks/Common/Shared/LogHelper.cs
@@ -5,23 +5,6 @@ using Microsoft.Crank.EventSources;
 
 namespace System.Net.Benchmarks;
 
-public static class MetricName
-{
-    public const string Avg = "/avg";
-    public const string Min = "/min";
-    public const string P50 = "/p50";
-    public const string P75 = "/p75";
-    public const string P90 = "/p90";
-    public const string P99 = "/p99";
-    public const string Max = "/max";
-    public const string Mean = "/mean";
-    public const string Read = "/read";
-    public const string Write = "/write";
-    public const string Rps = "/rps";
-    public const string Errors = "/errors";
-    public const string Handshake = "/handshake";
-}
-
 public static class LogHelper
 {
     private static readonly Dictionary<string, (string Description, string Format, bool IsLogged)> s_metrics = [];
