+++ src/System/Net/Benchmarks/Common/Shared/MetricName.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace System.Net.Benchmarks;
+
+public static class MetricName
+{
+    public const string Avg = "/avg";
+    public const string Min = "/min";
+    public const string P50 = "/p50";
+    public const string P75 = "/p75";
+    public const string P90 = "/p90";
+    public const string P99 = "/p99";
+    public const string Max = "/max";
+    public const string Mean = "/mean";
+    public const string Read = "/read";
+    public const string Write = "/write";
+    public const string Rps = "/rps";
+    public const string Errors = "/errors";
+    public const string Handshake = "/handshake";
+}
