+++ src/System/Net/Benchmarks/Common/Tls/Client/ClientCertSelectionType.cs
@@ -1,0 +1,11 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace System.Net.Benchmarks.Tls;
+
+public enum ClientCertSelectionType
+{
+    CertContext,
+    Collection,
+    Callback,
+}
