+++ src/System/Net/Benchmarks/Common/Tls/Shared/ApplicationProtocolConstants.cs
